Применяет критерий Б14.
